Lineage: remove circular deps between models.ts and VisGraphGenerator.ts
