+++ packages/components/src/internal/components/lineage/node/DetailsList.tsx
@@ -1,7 +1,6 @@
 import React, { Fragment, PureComponent, ReactNode } from 'react';
 
-import { LineageItemWithMetadata, LineageIOWithMetadata, LineageNode } from '../models';
-import { LineageNodeCollection } from '../vis/VisGraphGenerator';
+import { LineageNodeCollection, LineageItemWithMetadata, LineageIOWithMetadata, LineageNode } from '../models';
 import { DEFAULT_ICON_URL, getLineageNodeTitle } from '../utils';
 import { NodeInteractionConsumer, WithNodeInteraction } from '../actions';
 import { LineageDataLink } from '../LineageDataLink';
